refactor: moving the clock and compass codes to the huds folder with interface
